+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* **FIX**: Fix the drag math when the dialog is smaller than the default dimensions.
+* **FIX**: Fix the test asset loading issue in `lib/draggable_dialog.dart`.
+
 ## 0.0.2
 
 * **FIX**: Fix some formatting issues and add more complete description to the pubspec.yaml file.